Cap PERCENTILE_N/MEDIAN methods at 1M rows per partition
Implements a 1M row cap on PERCENTILE_N/MEDIAN methods as a safeguard
against runaway memory usage.

Without this check it's possible that the dist nodes on each
partition could put an effectively infinite number of doubles into
the underlying STL vector, opening the door for memory exhaustion.

The large table test has been updated to fill each of the available
partitions up with 1M rows, test that the functions work properly, and
then add 1 more row and make sure that the function aborts
as expected.